BackEnd/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn==0.27.0
websockets==12.0
python-multipart==0.0.6

# MongoDB
motor==3.3.2
pymongo[srv]==4.6.1

# Redis (if still using for other purposes)
redis==5.0.1
aioredis==2.0.1

# OpenAI
openai==1.6.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0

# Logging and async support
aiohttp==3.9.1
async-timeout==4.0.3

# Optional but recommended
ujson==5.9.0  # faster JSON processing

# Pin httpx to a version compatible with OpenAI library
httpx==0.27.2

# Audio processing
pydub==0.25.1
ffmpeg-python==0.2.0


PyPDF2==3.0.1  # or
pdfplumber==0.10.3  # Better for text extraction
python-docx==1.0.1  # Optional for Word docs

pyjwt==2.8.0  # Required for JWT handling
anthropic==0.25.4  # Required for Claude fallback

python-magic==0.4.27  # Required for file type detection

# Add this line for JWT cryptography support
cryptography==42.0.5

python-jose==3.3.0  # For JWT handling with cryptography support

SpeechRecognition==3.10.0